Major internal refactoring
Instead of the evolution matrices, the anomalous dimension matrices
are now directly implemented in the code. The evolution matrices are
then calculated numerically according to the usual formulae. Memoization
is used to ensure this does not have to be repeated when running multiple
times.

This new procedure makes the code cleaner, more maintainable, more stable
and more easily extensible.